requirements.txt
numpy==1.12.1
pypng==0.0.18
